기타사항

+++ lib/screen/client/main/tab/shorts/edit/s_video_editor.dart
@@ -80,7 +80,7 @@ class _VideoEditingScreenState extends ConsumerState<VideoEditingScreen> {
     await ExportService.runFFmpegCommand(
       await config.getExecuteConfig(),
       onProgress: (stats) {
-        _exportingProgress.value = config.getFFmpegProgress(stats.getTime());
+        _exportingProgress.value = config.getFFmpegProgress(stats.getTime().toInt());
       },
       onError: (e, s) => _showErrorSnackBar("Error on export video :("),
       onCompleted: (file) {
